Fix for the duplicate id flattenChildren warning
The unnecessary randomizing of the index in the getReposFromAPI
function was re-visiting objects that have already been added into the
return array. Added a dupesList object to keep track of the current
key/values added to the payload of the current request.  Will need to
figure out a better way to randomize/shuffle the github API.